+++ eras/conway/impl/test/Main.hs
@@ -6,7 +6,6 @@ module Main where
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Tx (tierRefScriptFee)
-import qualified Test.Cardano.Ledger.Alonzo.Imp as Alonzo
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Binary.CddlSpec as Cddl
 import qualified Test.Cardano.Ledger.Conway.GenesisSpec as Genesis
@@ -22,8 +21,8 @@ import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 instance EraSpec ConwayEra where
   eraImpSpec era = do
-    Alonzo.alonzoEraSpecificSpec era
-    Imp.conwayEraSpecificSpec
+    Imp.alonzoToConwaySpec era
+    Imp.conwayOnlySpec
     Imp.spec era
 
 main :: IO ()